WiP. A more generalized solution.
Still not ideal, but appears to work.
